requirements.txt
sklearn
numpy
pandas
matplotlib
nest_asyncio
python-telegram-bot